+++ lib/screens/explore_screens/regions_screens/regions_main/components/region_event_card/interactor/viewmodels/region_event_card_bloc.dart
@@ -2,7 +2,7 @@ import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:seeds/datasource/remote/model/profile_model.dart';
 import 'package:seeds/screens/explore_screens/regions_screens/regions_main/components/region_event_card/interactor/mappers/region_event_profiles_state_mapper.dart';
-import 'package:seeds/screens/explore_screens/regions_screens/regions_main/components/region_event_card/interactor/usecases/get_n_profiles_use_case.dart';
+import 'package:seeds/screens/explore_screens/regions_screens/regions_main/components/region_event_card/interactor/usecases/get_first_n_profiles_elements_use_case.dart';
 
 part 'region_event_card_event.dart';
 part 'region_event_card_state.dart';
@@ -14,8 +14,9 @@ class RegionEventCardBloc extends Bloc<RegionEventCardEvent, RegionEventCardStat
 
   Future<void> _onLoadRegionEventMembers(OnLoadRegionEventMembers event, Emitter<RegionEventCardState> emit) async {
     if (event.eventUsers.isNotEmpty) {
-      final results = await GetNProfilesUseCase().run(GetNProfilesUseCase.input(event.eventUsers, 3));
-      emit(RegionEventProfileMembersStateMapper().mapResultsToState(currentState: state, results: results));
+      final result =
+          await GetFirstNProfilesElementsUseCase().run(GetFirstNProfilesElementsUseCase.input(event.eventUsers, 3));
+      emit(RegionEventProfileMembersStateMapper().mapResultsToState(currentState: state, result: result));
     }
   }
 }
